Polyakov exp

@erkays @ss2959
1. Fixes a bug in perturbation sampling for larger bases
2. Adds support for even larger bases in pre-image sampling (the current limit is 2^22 -- 2^24, depending on the values of n and log2 q)

See merge request !274

+++ src/core/lib/lattice/field2n.h
@@ -72,7 +72,7 @@ public:
 	 * @brief Constructor from a ring element matrix
 	 * @param &element ring element matrix
 	 */
-	explicit Field2n(const Matrix<int32_t> & element);
+	explicit Field2n(const Matrix<int64_t> & element);
 
 	/**
 	 * @brief Method for getting the format/representation of the element
