+++ src/trapdoor/lib/sampling/dgsampling.h
@@ -107,7 +107,7 @@ public:
 	* @param *p non-spherical perturbation vector; output of the function
 	*/
 	static void ZSampleSigma2x2(const Field2n & a, const Field2n & b,
-		const Field2n & d, const Matrix<Field2n> &c, const typename Element::DggType & dgg, shared_ptr<Matrix<int32_t>> p);
+		const Field2n & d, const Matrix<Field2n> &c, const typename Element::DggType & dgg, shared_ptr<Matrix<int64_t>> p);
 
 	/**
 	* Subroutine used by ZSampleSigma2x2
@@ -117,7 +117,7 @@ public:
 	* @param dgg discrete Gaussian generator
 	* @param n ring dimension used for rejection sampling
 	*/
-	static shared_ptr<Matrix<int32_t>> ZSampleF(const Field2n &f, const Field2n &c,
+	static shared_ptr<Matrix<int64_t>> ZSampleF(const Field2n &f, const Field2n &c,
 		const typename Element::DggType &dgg, size_t n);
 
 private:
@@ -138,7 +138,7 @@ private:
 
 	//subroutine used by ZSampleF
 	//Algorithm utilizes the same inverse permutation algorithm as discussed in the GM17 paper
-	static void InversePermute(shared_ptr<Matrix<int32_t>> p);
+	static void InversePermute(shared_ptr<Matrix<int64_t>> p);
 
 };
